requirements.txt
Flask
Flask-CORS
pandas
scikit-learn
geopy
numpy
requests
python-dotenv
gunicorn